Single connection patch
Flyway now supports use of a single jdbc connection for metadata table management and
applying migrations. This feature is helpful in a situation where a migration takes
longer than the wait_timeout period. When this setting is set to false(default), migrations
that exceed the wait_timeout period will cause Flyway to crash due to the separate jdbc
connection for metadata table being shutdown.

The useSingleConnection property should be disabled for better performance.